requirement.txt
django
python-decouple
djangorestframework
djangorestframework-jwt
django-rest-passwordreset
pillow
###django-adminlte-3